Split evaluation benchmark harness into Bench.* modules.
Keep Main thin and group shared helpers, import-tree fixtures, substitution probes, and semisemantic probes for easier maintenance.
